+++ CHANGES.rst
@@ -11,7 +11,8 @@ Bugfix:
  -
 
 API Change:
- -
+ - Parameter historyVisibility removed from MxSession.createRoom(). It had no effect.
+ - New API: CreateRoomParams.setHistoryVisibility(String historyVisibility) to force the history visibility during Room creation.
 
 Translations:
  -